requirements.txt
streamlit
PyPDF2
python-dotenv
zhipuai
faiss-cpu
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-text-splitters==0.2.2